+++ tests/test_threads.py
@@ -1,11 +1,13 @@
 import threading
 import time
+from concurrent.futures import CancelledError
+from contextlib import suppress
 
 import pytest
 
 from anyio import (
     run_async_from_thread, run_sync_in_worker_thread, create_task_group, sleep,
-    create_capacity_limiter, create_event)
+    create_capacity_limiter, create_event, create_blocking_portal, start_blocking_portal)
 
 pytestmark = pytest.mark.anyio
 
@@ -130,3 +132,105 @@ async def test_cancel_worker_thread(cancellable, expected_last_active):
 
     await finish_event.wait()
     assert last_active == expected_last_active
+
+
+class TestBlockingPortal:
+    async def test_successful_call(self):
+        async def async_get_thread_id():
+            return threading.get_ident()
+
+        def external_thread():
+            thread_ids.append(portal.call(threading.get_ident))
+            thread_ids.append(portal.call(async_get_thread_id))
+
+        thread_ids = []
+        async with create_blocking_portal() as portal:
+            thread = threading.Thread(target=external_thread)
+            thread.start()
+            await run_sync_in_worker_thread(thread.join)
+
+        for thread_id in thread_ids:
+            assert thread_id == threading.get_ident()
+
+    async def test_aexit_with_exception(self):
+        """Test that when the portal exits with an exception, all tasks are cancelled."""
+        def external_thread():
+            try:
+                portal.call(sleep, 3)
+            except BaseException as exc:
+                results.append(exc)
+            else:
+                results.append(None)
+
+        results = []
+        with suppress(Exception):
+            async with create_blocking_portal() as portal:
+                thread1 = threading.Thread(target=external_thread)
+                thread1.start()
+                thread2 = threading.Thread(target=external_thread)
+                thread2.start()
+                await sleep(0.1)
+                assert not results
+                raise Exception
+
+        await run_sync_in_worker_thread(thread1.join)
+        await run_sync_in_worker_thread(thread2.join)
+
+        assert len(results) == 2
+        assert isinstance(results[0], CancelledError)
+        assert isinstance(results[1], CancelledError)
+
+    async def test_aexit_without_exception(self):
+        """Test that when the portal exits, it waits for all tasks to finish."""
+        def external_thread():
+            try:
+                portal.call(sleep, 0.2)
+            except BaseException as exc:
+                results.append(exc)
+            else:
+                results.append(None)
+
+        results = []
+        async with create_blocking_portal() as portal:
+            thread1 = threading.Thread(target=external_thread)
+            thread1.start()
+            thread2 = threading.Thread(target=external_thread)
+            thread2.start()
+            await sleep(0.1)
+            assert not results
+
+        await run_sync_in_worker_thread(thread1.join)
+        await run_sync_in_worker_thread(thread2.join)
+
+        assert results == [None, None]
+
+    async def test_call_portal_from_event_loop_thread(self):
+        async with create_blocking_portal() as portal:
+            exc = pytest.raises(RuntimeError, portal.call, threading.get_ident)
+            exc.match('This method cannot be called from the event loop thread')
+
+    @pytest.mark.parametrize('use_contextmanager', [False, True],
+                             ids=['contextmanager', 'startstop'])
+    def test_start_with_new_event_loop(self, anyio_backend_name, anyio_backend_options,
+                                       use_contextmanager):
+        async def async_get_thread_id():
+            return threading.get_ident()
+
+        if use_contextmanager:
+            with start_blocking_portal(anyio_backend_name, anyio_backend_options) as portal:
+                thread_id = portal.call(async_get_thread_id)
+        else:
+            portal = start_blocking_portal(anyio_backend_name, anyio_backend_options)
+            try:
+                thread_id = portal.call(async_get_thread_id)
+            finally:
+                portal.call(portal.stop)
+
+        assert isinstance(thread_id, int)
+        assert thread_id != threading.get_ident()
+
+    def test_call_stopped_portal(self, anyio_backend_name, anyio_backend_options):
+        portal = start_blocking_portal(anyio_backend_name, anyio_backend_options)
+        portal.call(portal.stop)
+        pytest.raises(RuntimeError, portal.call, threading.get_ident).\
+            match('This portal is not running')